+++ src/runtime/components/Error.vue
@@ -69,17 +69,17 @@ function handleError() {
 
 <template>
   <Primitive :as="as" data-slot="root" :class="b24ui.root({ class: [uiProp?.root, props.class] })">
-    <p v-if="!!props.error?.statusCode || !!slots.statusCode" data-slot="statusCode" :class="b24ui.statusCode({ class: uiProp?.statusCode })">
+    <p v-if="!!props.error?.statusCode || !!props.error?.status || !!slots.statusCode" data-slot="statusCode" :class="b24ui.statusCode({ class: uiProp?.statusCode })">
       <slot name="statusCode">
-        {{ props.error?.statusCode }}
+        {{ props.error?.statusCode || props.error?.status }}
       </slot>
     </p>
-    <h1 v-if="!!props.error?.statusMessage || !!slots.statusMessage" data-slot="statusMessage" :class="b24ui.statusMessage({ class: uiProp?.statusMessage })">
+    <h1 v-if="!!props.error?.statusMessage || !!props.error?.statusText || !!slots.statusMessage" data-slot="statusMessage" :class="b24ui.statusMessage({ class: uiProp?.statusMessage })">
       <slot name="statusMessage">
-        {{ props.error?.statusMessage }}
+        {{ props.error?.statusMessage || props.error?.statusText }}
       </slot>
     </h1>
-    <p v-if="(props.error?.message && props.error.message !== props.error.statusMessage) || !!slots.message" data-slot="message" :class="b24ui.message({ class: uiProp?.message })">
+    <p v-if="(props.error?.message && props.error.message !== (props.error.statusMessage || props.error.statusText)) || !!slots.message" data-slot="message" :class="b24ui.message({ class: uiProp?.message })">
       <slot name="message">
         {{ props.error?.message }}
       </slot>